+++ react-app/src/Hooks/GetAll.tsx
@@ -1,24 +1,24 @@
-import { useState } from 'react';
-import axios from 'axios';
+// import { useState } from 'react';
+// import axios from 'axios';
 
-const GetAllOrders = (): [() => Promise<void>, any, boolean, any] => {
-    const url: string = `${process.env.URL}/api/orders`;
+// const GetAllOrders = (): [() => Promise<void>, any, boolean, any] => {
+//     const url: string = `${process.env.URL}/api/orders`;
 
-    const [res, setRes] = useState<any>(null);
-    const [loading, setLoading] = useState<boolean>(false);
-    const [error, setError] = useState<any>(null);
+//     const [res, setRes] = useState<any>(null);
+//     const [loading, setLoading] = useState<boolean>(false);
+//     const [error, setError] = useState<any>(null);
    
-    const get = async (): Promise<void> => {
-        setLoading(true);
-        try { 
-            const response = await axios.get(url);
-            setRes(response.data);
-        } catch (error) { 
-            setError(error);
-        } finally {
-            setLoading(false);
-        }
-    };  
-    return [get, res, loading, error];
-};
-export default GetAllOrders;
+//     const get = async (): Promise<void> => {
+//         setLoading(true);
+//         try { 
+//             const response = await axios.get(url);
+//             setRes(response.data);
+//         } catch (error) { 
+//             setError(error);
+//         } finally {
+//             setLoading(false);
+//         }
+//     };  
+//     return [get, res, loading, error];
+// };
+// export default GetAllOrders;
