+++ react-app/src/Hooks/Post.tsx
@@ -1,14 +1,14 @@
-import axios from "axios";
+// import axios from "axios";
 
-const AddAnOrder = async (order: any): Promise<any> => {
-    const url: string = `${process.env.URL}/api/orders`;
-    try {
-        const response = await axios.post(url, order);
-        return response.data;
-    } catch (error) {
-        console.error("Error adding order:", error);
-        throw error;
-    }
-};
+// const AddAnOrder = async (order: any): Promise<any> => {
+//     const url: string = `${process.env.URL}/api/orders`;
+//     try {
+//         const response = await axios.post(url, order);
+//         return response.data;
+//     } catch (error) {
+//         console.error("Error adding order:", error);
+//         throw error;
+//     }
+// };
 
-export default AddAnOrder;
+// export default AddAnOrder;
